Rename auto registration module to registration
