+++ applications/tari_validator_node/src/bootstrap.rs
@@ -39,7 +39,7 @@ use tari_p2p::initialization::spawn_comms_using_transport;
 use tari_shutdown::ShutdownSignal;
 
 use crate::{
-    auto_registration,
+    registration,
     base_layer_scanner,
     comms,
     grpc::services::base_node_client::GrpcBaseNodeClient,
@@ -167,7 +167,7 @@ pub async fn spawn_services(
     save_identities(config, &comms)?;
 
     // Auto-registration
-    auto_registration::spawn(config.clone(), node_identity.clone(), epoch_manager.clone(), shutdown);
+    registration::spawn(config.clone(), node_identity.clone(), epoch_manager.clone(), shutdown);
 
     Ok(Services {
         comms,
