+++ applications/tari_validator_node/src/json_rpc/handlers.rs
@@ -47,7 +47,7 @@ use tari_validator_node_client::types::{
 };
 
 use crate::{
-    auto_registration,
+    registration,
     grpc::services::{base_node_client::GrpcBaseNodeClient, wallet_client::GrpcWalletClient},
     json_rpc::jrpc_errors::internal_error,
     p2p::services::{
@@ -140,7 +140,7 @@ impl JsonRpcHandlers {
     pub async fn register_validator_node(&self, value: JsonRpcExtractor) -> JrpcResult {
         let answer_id = value.get_answer_id();
 
-        let resp = auto_registration::register(self.wallet_client(), self.node_identity.clone(), &self.epoch_manager)
+        let resp = registration::register(self.wallet_client(), self.node_identity.clone(), &self.epoch_manager)
             .await
             .map_err(internal_error(answer_id))?;
