+++ applications/tari_validator_node/src/registration.rs
